requirements.txt
SpeechRecognition
gTTS
playsound==1.2.2
rich
google-generativeai
# openai
PyAudio
dotenv